adds a lot more adjacencies
Actually, this commit adds _a lot_ of adjacencies that might be of
limited use, so the exact heuristic for determining which adjacencies to
include or not is still to be refined. Bear in mind that this commit
increases compilation times, significantly, as well as the amount of
overrides in the output.

Besides, this commit adds distinct Sets and Seqs for the override
networks to overcome limitations in Scala's Enumeration implementation.
An update introducing my own adaption of Enums to overcome this problem
is pending.

Moreover, a few functions related to properties of the networks have
been separated from various places into a common module to keep future
modifications simple and in one place.