decouple the definition of networks from metarules implementation
This uses dependency injection to let the calling program downstream
supply the concrete definition of networks, so that we do not need to
hardcode the networks here. This way the metarules package becomes more
independent from a concrete choice of implemented networks, so end users
do not need to modify this library as often anymore.

The implementation wraps parts of the metarules library in a `Syntax`
trait with a `Network` type that is provided downstream.